refactor(product): split oversized viewer, coordinator, and host files
Line-cap-only code motion inside the owning 02A workstreams:
FileViewerToolbar extracted from FileViewerFrame, the persistence
read/write cycle extracted from the dock coordinator, and ProductEntry
types extracted from product-host (plus doc tightening). Removes the
now-stale product-host frontend-structure ratchet entry (file dropped
below the base threshold; the report tool fails on stale entries).
