+++ apps/packages/product-client/src/hooks/app/lifecycle/use-product-entry-routing.ts
@@ -1,7 +1,7 @@
 import { useEffect, useRef } from "react";
 import { useNavigate } from "react-router-dom";
 import { useProductHost } from "@proliferate/product-client/host/ProductHostProvider";
-import type { ProductEntry } from "@proliferate/product-client/host/product-host";
+import type { ProductEntry } from "@proliferate/product-client/host/product-host-entry";
 import { productEntryRoute } from "#product/lib/domain/auth/desktop-navigation";
 
 /**
